remove stack chart advisor
