+++ src/SmartComponents/Vulnerability/VulnerabilityCard.js
@@ -54,8 +54,6 @@ class VulnerabilityCard extends Component {
             { name: 'CVSS 4.0 - 7.9', x: '2015', y: 20, fill: '#ec7a08' },
             { name: 'CVSS 0.0 - 3.9', x: '2015', y: 40, fill: '#f0ab00' }
         ];
-        const stackChartPadding = { bottom: 0, left: 0, right: 0, top: 0 };
-
         const stackChartLegendData = chartData.map(item => ({ name: `${item.name}`, symbol: { fill: `${item.fill}`, type: 'circle' } }));
 
         return (
@@ -92,8 +90,6 @@ class VulnerabilityCard extends Component {
                     <StackChartTemplate
                         ariaDesc="CVEs impacting your systems"
                         ariaTitle="Vulnerabilities chart"
-                        padding={ stackChartPadding }
-                        legendPosition="bottom-left"
                         height={ 40 }
                         width={ 600 }
                         maxWidth={ 600 }
